Rename AssignableRoles to have .dto.ts suffix
Missed in a3d88c70c25833c01f5943b75728460aa9a69d1b
